+++ src/startgg/utils.ts
@@ -1,12 +1,13 @@
-import { iSet } from "./types.ts";
+import { sggSet } from "./types.ts";
 
-export const isSetDefined = (set: iSet) =>
+export const isSetDefined = (set: sggSet) =>
   set.slots[0].entrant && set.slots[1].entrant;
 
-export const isSetUnstarted = (set: iSet) =>
+export const isSetUnstarted = (set: sggSet) =>
   set.state === 1 && isSetDefined(set);
 
-export const isSetPlaying = (set: iSet) => set.state === 2 && isSetDefined(set);
+export const isSetPlaying = (set: sggSet) =>
+  set.state === 2 && isSetDefined(set);
 
-export const isSetFinished = (set: iSet) =>
+export const isSetFinished = (set: sggSet) =>
   set.state === 3 && isSetDefined(set);